requirements.txt
Flask
requests
Pillow
flask_cors